fix: 일정 수정시 화면상 날짜가 바뀌지 않는 문제 개선
- formattedDate에 따라 ModalDate를 다르게 호출하는 부분을 통일하고, 상위 컴포넌트에서 formattedDate로 뿌려주는 방식으로 변경
